requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
mysql-connector==2.1.6
mysqlclient==1.3.14
pkg-resources==0.0.0
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1

